[update]component側で行っていたリクエストの処理をControllerに委嘱
